Refactor: some renames

+++ src/kernel/lapic.c
@@ -1,4 +1,4 @@
 #include "rt64.h"
 
-// detected and initialized in acpi
-volatile usize *lapic;
+// detected and initialized in acpi.c, points to physical addr 0xFEE0_0000
+volatile u32 *lapic;
